requirements.txt
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
colorama==0.4.6
cycler==0.11.0
exceptiongroup==1.2.2
fastapi==0.103.2
ffmpy==0.3.2
filelock==3.12.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
fvcore==0.1.5.post20221221
gradio==3.34.0
gradio_client==0.2.6
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.10
Jinja2==3.1.6
jsonschema==4.17.3
kiwisolver==1.4.5
linkify-it-py==2.0.3
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mmcv-full==1.3.0
multidict==6.0.5
numpy==1.21.6
omegaconf==2.3.0
opencv-python==4.8.0.76
orjson==3.9.7
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
portalocker==2.7.0
pycocotools==2.0.6
pydantic==1.10.7
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.4
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-multipart==0.0.8
pytz==2025.2
pywin32==308
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
semantic-version==2.10.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tabulate==0.9.0
termcolor==2.3.0
terminaltables==3.1.10
timm==0.9.5
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.1
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.67.1
transformers==4.30.2
typing_extensions==4.7.1
uc-micro-py==1.0.3
urllib3==2.0.7
uvicorn==0.22.0
websockets==11.0.3
Werkzeug==2.2.3
wincertstore==0.2
yacs==0.1.8
yapf==0.43.0
yarl==1.9.4
zipp==3.15.0
